rename bdiff.c to isequal.c, make it handle different line-endings as equal
